Before #5796, we were using F16 KV cache. Cast K and V to F16 to match the performance