@slaren 2024-03-05
I noticed that the first evaluation with BERT is a lot faster than the rest (about 2x), hence the huge variance. Is this expected? Maybe there is some cleanup necessary between runs that `llama-bench` is not doing?

```
$ ./llama-bench -m models/nomic-embed/ggml-model-f16.gguf -p  8192 -n 0 -o json -r 20
...
    "samples_ns": [ 55305409, 106761860, 103784178, 103219933, 102769295, 106150938, 103377148, 103227991, 103466474, 103412610, 103204537, 103091448, 102517562, 102943769, 102898136, 102444703, 103308155, 103430029, 103645031, 103997954 ],
    "samples_ts": [ 148123, 76731.5, 78933, 79364.5, 79712.5, 77173.1, 79243.8, 79358.3, 79175.4, 79216.6, 79376.4, 79463.4, 79908.3, 79577.4, 79612.7, 79965.1, 79296.7, 79203.3, 79039, 78770.8 ]
  }
```