@ggerganov 2024-03-05
Similar observation on my GPU:

```
  Device 0: NVIDIA GeForce RTX 2060 SUPER, compute capability 7.5, VMM: yes
[
  {
    "build_commit": "40ca2e03",
    "build_number": 2351,
    "model_type": "bert 33M F16",
    ...
    "samples_ns": [ 132941728, 147430247, 147204893, 147173373, 147210412, 147196176, 147225462, 147165247, 147198780, 147198841 ],
    "samples_ts": [ 61621, 55565.3, 55650.3, 55662.2, 55648.2, 55653.6, 55642.5, 55665.3, 55652.6, 55652.6 ]
  }
]
```

Not expected. Maybe some issue with the input preparation, though I'm not sure where it could be