@slaren 2024-03-05
The problem is that `llama-bench` does not set `cparams.embeddings`, and the model does not produce logits either. The result is that there isn't any data copied back to the CPU from the model, and there aren't any calls made to `ggml_backend_synchronize`. So all that it is doing is queueing evaluations on the GPU and never waiting for them to end, since `ggml_backend_graph_compute` in the CUDA backend is asynchronous. Setting `cparams.embeddings` to `true` in `llama-bench` results in much more repeatable (and reasonable) results:
```
"samples_ts": [ 15884.9, 17033.2, 17236.4, 17281.7, 17280.4, 17326.7, 17305.6, 17287.7, 17252.3, 17251.9, 17300.1, 17247.9, 17218.3, 17208.6, 17266.2, 17175.7, 17189.4, 17236.7, 17301.3, 17290.8 ]
```

With that fix applied, and the fix to `compare-llama-bench.py`, I get these results vs master:

```
scripts/compare-commits.sh master gg/bert-f16  -m models/nomic-embed/ggml-model-f16.gguf -p 512,1024,2048,4096,8192 -b 8192 -n 0 -r 10           
```

| GPU         | Model               | Test   |   t/s master |   t/s gg/bert-f16 |   Speedup |
|:------------|:--------------------|:-------|-------------:|------------------:|----------:|
| RTX 3090 Ti | nomic-bert 137M F16 | pp512  |     55166.66 |          72500.62 |      1.31 |
| RTX 3090 Ti | nomic-bert 137M F16 | pp1024 |     52499.25 |          61251.00 |      1.17 |
| RTX 3090 Ti | nomic-bert 137M F16 | pp2048 |     48672.47 |          48633.91 |      1.00 |
| RTX 3090 Ti | nomic-bert 137M F16 | pp4096 |     33049.97 |          30725.15 |      0.93 |
| RTX 3090 Ti | nomic-bert 137M F16 | pp8192 |     20921.86 |          17152.72 |      0.82 |