@ggerganov 2024-03-06
It's surprising that for large batch sizes the F32 version is faster. On RTX 2060 with 33M model, F16 is faster only for BS = 512:

- `master`

| model                          |       size |     params | backend    | ngl |    n_batch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 512     | 45000.11 ± 3151.59 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 1024    | 35720.98 ± 128.35 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 2048    | 28716.76 ± 121.26 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 4096    | 19417.52 ± 49.16 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 8192    |   9123.49 ± 6.09 |

build: bd836944 (2350)

- PR

| model                          |       size |     params | backend    | ngl |    n_batch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 512     | 50991.57 ± 365.12 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 1024    | 35353.14 ± 40.56 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 2048    | 26109.02 ± 33.45 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 4096    |  16230.26 ± 6.94 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 8192    |   8297.65 ± 2.88 |

build: 40ca2e03 (2351)

- pre #5796 

| model                          |       size |     params | backend    | ngl |    n_batch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 512     | 50444.46 ± 387.10 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 1024    | 34794.17 ± 27.87 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 2048    | 25528.46 ± 13.41 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 4096    |  15831.46 ± 7.02 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 8192    |   8030.40 ± 6.07 |

build: e0843afe (2340)

Should we keep the `master` version?
I imaging that embeddings are usually computed on large amounts of tokens and/or many small texts batched together. So large batch sizes seem more important than small ones