@iamlemec 2024-03-06
Interesting. I'm seeing across the board speedups on an RTX A6000. Here I did a bunch of reps to average out the first run issues. Looks pretty close to the theoretical 2x for large batches. I wonder if the 2060 is not using the tensor cores for FP16 for some reason? The numbers look a lot like you're paying a small cost for the conversion with no intrinsic speedup.

One use case for small batches is when you're trying to reduce response time for a single query to a vector db. Those will typically be small and single sequence. For batched small-ish chunks, I think you may as well keep batch sizes small to avoid quadratic attention costs. I guess the newer long context stuff would be the concern. Might be nice to see how it looks on an RTX 30 or 40 series.

Master
| model                          |       size |     params | backend    | ngl |    n_batch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 512     | 105369.50 ± 21810.07 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 1024    | 106351.34 ± 48206.62 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 2048    | 85112.62 ± 49937.49 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 4096    | 57543.13 ± 57965.89 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 8192    | 30851.35 ± 30181.20 |

PR
| model                          |       size |     params | backend    | ngl |    n_batch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 512     | 148597.59 ± 61330.15 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 1024    | 133139.99 ± 136358.94 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 2048    | 102929.04 ± 123899.19 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 4096    | 53393.35 ± 64028.78 |

Pre fix-embeddings (e0843afe):
| model                          |       size |     params | backend    | ngl |    n_batch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 512     | 148665.81 ± 57567.63 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 1024    | 124998.71 ± 98632.21 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 2048    | 91358.75 ± 88137.55 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 4096    | 44572.14 ± 41171.58 |
| bert 33M F16                   |  63.46 MiB |    33.21 M | CUDA       |  99 |       8192 | pp 8192    | 25118.04 ± 24507.47 |